refactor: extract hook to separate file
