+++ src/components/NucleotideEntropy/NucleotideEntropyMultiChart.tsx
@@ -1,7 +1,5 @@
-import React, { ChangeEvent, Dispatch, SetStateAction, useMemo, useState } from 'react';
-import { MutationProportionData } from '../../data/MutationProportionDataset';
+import React, { ChangeEvent, Dispatch, SetStateAction, useState } from 'react';
 import Loader from '../Loader';
-import { useQuery } from '../../helpers/query-hook';
 import { LapisSelector } from '../../data/LapisSelector';
 import { PipeDividedOptionsButtons } from '../../helpers/ui';
 import { Legend, Line, LineChart, ResponsiveContainer, Tooltip, XAxis, YAxis } from 'recharts';
@@ -10,16 +8,13 @@ import FormGroup from '@mui/material/FormGroup';
 import FormControlLabel from '@mui/material/FormControlLabel';
 import { SequenceType } from '../../data/SequenceType';
 import { Form } from 'react-bootstrap';
-import { globalDateCache } from '../../helpers/date-cache';
-import { FixedDateRangeSelector } from '../../data/DateRangeSelector';
-import { DateRange } from '../../data/DateRange';
 import jsonRefData from '../../data/refData.json';
-import { GeneOption, weeklyMeanEntropy } from './calculateEntropy';
+import { GeneOption } from './calculateEntropy';
 import { useExploreUrl } from '../../helpers/explore-url';
 import { pprettyColors } from '../../helpers/colors';
-import { getTicks } from '../../helpers/ticks';
 import { formatDate } from '../../widgets/VariantTimeDistributionLineChartInner';
 import { genes } from './genes';
+import { useMultiChartData } from './useMultiChartData';
 
 type Props = {
   selectors: LapisSelector[];
@@ -130,98 +125,6 @@ function Controls(props: {
   );
 }
 
-const useData = (
-  selectors: LapisSelector[],
-  selectedGene: GeneOption[],
-  variants: string[],
-  sequenceType: SequenceType,
-  deletions: boolean
-) => {
-  const days = [
-    selectors[0].dateRange?.getDateRange().dateFrom!,
-    selectors[0].dateRange?.getDateRange().dateTo!,
-  ];
-  const dayRange = globalDateCache.rangeFromDays(days)!;
-  const weekDateRanges: DateRange[] = globalDateCache
-    .weeksFromRange({
-      min: dayRange.min.isoWeek,
-      max: dayRange.max.isoWeek,
-    })
-    .map(week => ({
-      dateFrom: week.firstDay,
-      dateTo: week.firstDay,
-    }));
-
-  const weekSelectors = selectors.flatMap(selector =>
-    weekDateRanges.map(weekRange => ({
-      ...selector,
-      dateRange: new FixedDateRangeSelector(weekRange),
-    }))
-  );
-
-  let weekRangesCount = weekDateRanges.length;
-
-  const weeklyVariantMutationProportionQuery = useQuery(
-    async signal =>
-      await Promise.all(
-        weekSelectors.map((weekSelector, i) =>
-          MutationProportionData.fromApi(weekSelector, sequenceType, signal).then(data => ({
-            proportions: data.payload,
-            date: weekDateRanges[i % weekRangesCount],
-          }))
-        )
-      ),
-    [selectors, sequenceType]
-  );
-
-  return useMemo(() => {
-    if (!weeklyVariantMutationProportionQuery.data) {
-      return undefined;
-    }
-    if (!selectedGene) {
-      return undefined;
-    }
-    if (!variants[0]) {
-      return undefined;
-    }
-    if (variants.length * weekRangesCount !== weeklyVariantMutationProportionQuery.data.length) {
-      return undefined;
-    }
-
-    const dates = weeklyVariantMutationProportionQuery.data.map(proportionData => {
-      return { date: proportionData.date.dateFrom?.dayjs.toDate()! };
-    });
-    const ticks = getTicks(dates);
-
-    const weeklyDataByTimestamp = weeklyMeanEntropy(
-      weeklyVariantMutationProportionQuery.data,
-      sequenceType,
-      selectedGene[0],
-      deletions
-    )
-      .map(({ week, meanEntropy }, i) => ({
-        day: week.dateFrom!.dayjs.toDate().getTime(),
-        [variants[Math.floor(i / weekRangesCount)]]: meanEntropy,
-      }))
-      .reduce((aggregated, weeklyMeanEntropy) => {
-        const previousValue = aggregated[weeklyMeanEntropy.day] ?? {};
-        aggregated[weeklyMeanEntropy.day] = { ...previousValue, ...weeklyMeanEntropy };
-        return aggregated;
-      }, {} as Record<number, any>);
-
-    let plotData = Object.values(weeklyDataByTimestamp); //depending on the day, the latest week just started, so the entropy is calculated as 0 because there are no samples
-
-    return { plotData, ticks };
-  }, [
-    weeklyVariantMutationProportionQuery,
-    selectedGene,
-    variants,
-    deletions,
-    sequenceType,
-    weekRangesCount,
-  ]);
-};
-
 type PlotProps = {
   selectors: LapisSelector[];
   selectedGeneOptions: GeneOption[];
@@ -231,7 +134,7 @@ type PlotProps = {
 };
 
 const Plot = ({ selectors, selectedGeneOptions, variants, sequenceType, includeDeletions }: PlotProps) => {
-  const data = useData(selectors, selectedGeneOptions, variants, sequenceType, includeDeletions);
+  const data = useMultiChartData(selectors, selectedGeneOptions, variants, sequenceType, includeDeletions);
 
   if (variants![0].length === 0) {
     return <p>Select one or more lineages to compare.</p>;
